+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- log: [logrus](https://github.com/sirupsen/logrus) adapter
+
 
 ## [0.20.0] - 2018-03-18